requirements.txt
flake8==3.8.4
ipython==7.19.0
lxml==4.6.2
matplotlib==3.3.3
pandas==1.2.0
pytest==6.2.1
python-dotenv==0.15.0
requests==2.25.1
